Bug 369922 - Bugzilla Adapter: Selection dialog not working
The Bugzilla selection dialog is not working after the upgrade to
j2bugzilla 1.0. We had some code in the ExtendedBugSearch class to
workaround a problem with searching in old versions of j2bugzilla, but
it's no longer needed (and is now causing problems). The fix is simply
to remove the workaround.